echo-bridge/requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
